Import SC15/Chameleon version of NRM
The SC15/Chameleon experiments were based on a simple power management
scheme built with the help of beacon for transport, RAPL for monitoring,
and socket-based communications between the NRM and the argobots
runtime.

This is an import of the working code we had, in the state it was at the
time. It is quite obvious that this kind of one time hack, and it
probably doesn't work without the exact Chameleon setup.