[refactor] Python rewrite of the software
We chose to rewrite the entire thing in python. The language should make
it easy to interact will all the moving parts of the Argo landscape, and
easy to prototype various control schemes.

The communication protocol is exactly the same, but implemented with
ZeroMQ + tornado.

Power readings are not integrated yet, we are targeting using the Coolr
project for that.

This is a rough draft, all the code is in binary scripts instead of
the package, and there are no unit tests. Nevertheless, it should be
a decent starting point for future development.