[refactor] Rename clients to applications
Giving the deadlines we have, it's probably a good idea to keep things
simple and keep the application facing protocol exactly as it is, and
restrict it to just applications. This way we can keep using the argobots
tests as valid benchmarks, and at the same start building a decent
communication protocol on a different socket, with a better interface.

To clarify that, the daemon now use the word application to refer to
clients connecting on the "legacy" interface.

We'll add a different socket and start building a real protocol in
future commits.