Import coolr files into project, link to sensor
This a partial import of github.com/coolr-hpc/pycoolr from
master branch, commit id: 67e7aa4b89b67744922b5926cd1459adf650013b

Coolr will provide us the capability to read power, topology and
msr-based sensors.

This patch links the sensor module with the coolr code as it is, which
doesn't really work. The core issue is that the current coolr code is
meant to be stand-alone, and the main functions to both sampling and
formatting of the data in json, by hand.

We will solve that in the next commit, removing the json-specific code
from the sample function to create a dict of values instead.