[refactor] Remove json formatting from clr_rapl
clr_rapl was merging sampling data generation and formatting that same
data into a json string. This patch removes the json formatting, to let
users of the module use the data in a python structure.

The patch also removes __main__ code from clr_rapl, as it is unnecessary
here. We might end up reusing some of this code in unit tests later, but
not right now.