[refactor] Use actual sensor info in control loop
This patch is the last link between coolr and the daemon. We now create
a dictionary of machine information inside the sensor manager, and give
this dictionary back to the daemon. The daemon can then use the real
data for control.

We still need to receive the target power from somewhere, and that will
come later.