[feature] Implement basic CLI
Only supports setpower for now, and while it should work in theory, the
current code doesn't have a way to check if the command was received, as
the daemon never advertise the current limit.

We need to change the protocol at this point.

This also fixes a bug in the daemon code, that was expecting a single
string as a message, instead of a list of parts, as zmqstream always
receives.