[feature] Add container run skeleton
This is the first step in a series of patches to integrate the container
launching code from Argus (globalos) into the NRM infrastructure.

This patch creates a valid command on the CLI, and sends the necessary
info to the NRM. We still need to take care of the actual container
creation.

Note that the CLI waits for an event indicating that the container was
launched, at that at this point the event is never generated by the NRM.