[feature] Add command to list containers
This patch adds a very simple command to list the containers currently
known by the NRM. There's no history or state tracking on the NRM, so
the code is pretty simple.

We expect that some of the container tracking doesn't need to be sent
for such a command, so the listing also filters some of the fields.

This patch also adds an 'event' field to container messages, as it would
probably be needed further for other kind of operations.