[test] Add a clr_rapl test matching current code
This patch adds a test for the nrm/coolr/clr_rapl.py code, based on
the __main__ code added to it recently. This is an example of how the
modules should be tested automatically, instead of relying on __main__
snippets everywhere.

We should replace those snippets bits by bits by tests similar to this
one. The end goal is to able to active a gitlab pipeline that validate
merge requests automatically.