[feature] Replace client with dummy application
This patch replace the client code (bin/client and nrm/client) by a new
application code that integrates progress reports and uses the new
downstream API.

While git is reporting that both codes are different, the app code is
basically a refactoring and adaptation of the client code.

This is directly related to issue #2.