[fix] Lock with Pipenv all packages
Some packages were being pulled into the virtual environment through
setup.py. It just happens that the new versions of pyzmq and tornado
don't really play nice with the current daemon code. Mostly because of a
"better" ioloop hijacking mecanism, that doesn't work for us.

This patch moves the install requirements of the setup.py to Pipfile,
and lock them to a working version.

Note: the code still currently triggers #14.