[feature/fix] add listen command for upstream pub
Add a listen command to get access to the event stream of the upstream
pub/sub API.

This patch gives back access from the command line to the power
information of a container, including filtering the event stream to only
have events relevent to this container.

This changes the workflow a little bit for users, but should result in a
cleaner access to profiling data in the future.

Related to #18.