[feature] add downstream rpc to messaging layer
Add downstream RPC client/server classes that are the same as the
upstream ones.

This is part of a series of changes to downstream to allow for more
reliable communications between the daemon and applications. At this
time, the daemon never replies, so the RPC_REQ is basically used as a
way to publish events to the daemon.