[ci] Pinning integration tests
This commit pins the integration tests. This is done in the .integration.nix
file, which in turns uses the derivations from the 'argotest'
repository. These pin argopkgs - which itself pins nixpkgs, so
everything should be completely reproducible.

The integration target provisions the dependencies using nix-shell and
runs the basic 'hello world' container test.