[refactor] use singularity instances
Instances in singularity are basically the same as being able to create
an empty container and then run something in it, so we can have the same
thing as with nodeos.

One main issue with the code as is: we should probably refactor the
container_manager code to modify the manifest on the fly according to
the options listed in it. It would make it easy to pass along stuff like
PWD, necessary binds and so on.